+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.1-wip
+
+- Add support on content generating methods for overriding "tools" passed when
+  the generative model was instantiated.
+- Add support for forcing the model to use or not use function calls to generate
+  content.
+
 ## 0.3.0
 
 - Allow specifying an API version in a `requestOptions` argument when